Setup VCR for testing
It's now as simple as decorating a test with `@pytest.mark.vcr()` to
enable VCR and make it record and replay interactions with FreeIPA.

The first time the test is run, it will interact with FreeIPA and change
its database. There is no method in this commit to reset the database to
a previous known state.

This commit also adds a test for the registration endpoint to
demonstrate VCR.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>