Don't delete the user when registration fails
This will avoid potential attack vectors, as well as actually work in production where we won't be allowed to delete users anyway.

When a password does not comply with policy, it will stay expired and the newly registered user will be sent to the login page where they will have to change it.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>