add the ability in the UI to add new GPG Key Ids
Adds the ability to add 2 additional GPG Key IDs on every
save of the edit user dialog. This basic solution was chosen
so we can use append_field() from WTForms fieldlists, rather
than some hand-crafted hacky javascript to add new fields on
the fly.

The drawback with this approach is that a user can only add 2
new GPG Keys at a time. More than 2 keys can be added, the user
just has to submit the form, and then add more.

Resolves: #96

Signed-off-by: Ryan Lerch <rlerch@redhat.com>