show membermanagers that arent members in group page
Previously, if a user was a membermanager (i.e. a sponsor) of
a group, but not a member of the group, they would not show up
in the list of membermanagers (sponsors) on the group listing
page. However, if that user was logged in and looking at the
group page, they still could edit the other members of the group.

This commit corrects the group page to show membermanagers(sponsors)
that happen to not be members of the group.

Related: #108

Signed-off-by: Ryan Lerch <rlerch@redhat.com>