make the otp_disable and otp_delete tests more real
The test_user_settings_otp_disable and
test_user_settings_otp_delete tests add two otp tokens, for
the purposes of removing (or disabling) one of them. However,
after adding the first token, to add the second token, the OTP would
have had to be supplied with the password. This was previously worked
around by the VCR cassette being generated on a system without OTP
enabled.

This PR updates the cassettes and checks that the OTP actaully works.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>