make loggedin change password form otp aware
previously, the form a logged in user used to change their
password presented fields and descriptions for OTP stuff
when a user hadnt enrolled an OTP token.

This hides those descriptions and fields if a user has not
enabled a OTP token.

Note that the descriptions are still shown on the expired password
reset page as this is not a logged in page, and we cannot know if the
user has an OTP setup there (nor would we want to expose that fact to
the world if not logged in)

Resolves: #206

Signed-off-by: Ryan Lerch <rlerch@redhat.com>