escape the OTP token description in the URI
Previously, code was added to noggin that edited the URI we got back
from freeipa, adding the description of the token to the OTP URI. this
was done, so the token description is shown in the authenticator
application that the user uses. However, if a user added a token
description with a space in it,  the URI would be constructed to be
invalid, and not be able to be scanned via the QR code.

This PR escapes the description before adding it to the otp URI, which
resolved this issue.

Resolves: #241

Signed-off-by: Ryan Lerch <rlerch@redhat.com>