Dynamically create the app object
Instead of instanciating the app as a global object, use the
[application factory](https://flask.palletsprojects.com/en/1.1.x/patterns/appfactories/)
pattern as recommended by Flask.

This means we have a root blueprint and register the controllers with it
rather than the main app, and as a consequence we need to prefix all
endpoints with a dot, which is what causes most of the changes in the
templates.

Nice side effect: we don't need to pass the configuration environment
variable just to run the unit tests anymore, which make them more
VSCode-friendly. We also don't have to set that environment variable to
build the docs or run `towncrier` either.

Also, reorganize some of the code to have a clearer separation of module
topics.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>