Set a BASE_URL variable in the master template to be used in search.js
Basically this variable allows to specify the full URL where noggin
is deployed. Then in search.js we are able to build the full URL
linking to where the user/group are.

The current behavior assumed that noggin was running at the root of
the web-server (ie: /) while this is not necessarily true (it could
run at say /accounts/).

This commit should fix that.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>