Add an optional note to the user settings OTP page
Add an optional note to the user settings OTP page that is defined in
the main.html in the themes templates to display a note at the top of
the OTP page.

This is used in Fedora Accounts to point users at the documentation for
getting a kerberos ticket when OTP is enabled.

Resolves: #344

Signed-off-by: Ryan Lerch <rlerch@redhat.com>