Fix multiple otp notices
The note added to the user's otp settings page in
de1bc0c56ff6c5aedb36847b9854b287db7262bc would be presented to the
user multiple times if more than one OTP token was defined.

This fixes this issue so it only shows once if at least one token
is defined, and not at all if no tokens are defined.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>