Update search with loading animations
Previously, there were no loading animations for the search box, and due
to the fact we are doing two seperate calls (for users and groups),
these could return at different times. So this commit adds sepearte
spinner animations for the two results classes.

Additionally, this moves the string for "users" and "groups" that is
used to label the groups in the search dropdown into the templates as
variables, so they can be translated.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>