Add optional note above the login / register box
This adds an additional macro that appears above the login / register
box on the unlogged in main page. The FAS theme also now uses this box
to inform users that if they use a Fedora account, they must adhere to
the Fedora Code Of Conduct.

This also slightly tweaks the links for sync token and forgotten
password to clean them up a littel bit.

Resolves: #495

Signed-off-by: Ryan Lerch <rlerch@redhat.com>