Default locale and timezone fallbacks in the UI
In some cases, a user may have been imported into ipa without a locale
or a timezone set. As these are required fields on the noggin side of
things, when they edit their profile for the first time, the <select>
boxes showed the first items in this list (af-ZA for locale and
Africa/Abidjan for Timezone.)

This commit fixes this issue by putting en-US and UTC at the top of the
lists of Locales and Timezones, so if these are not set, the select
boxes default to those values when editing the form.

Additionally, the UI for the select boxes now use selectize, so the user
can search for a locale and a timezone by substring. Selectize also
sorts the options when presenting them to the user as well.

https://pagure.io/fedora-infrastructure/issue/9761

Signed-off-by: Ryan Lerch <rlerch@redhat.com>