Restrict typer to ^0.11.0
rstcheck depends on  "typer[all]  >=0.4.1" , however in version 0.12.0
of typer, the `all` extra was renamed to `standard`, and this was
causing issues when running the license checks.

This commit adds an explicit limitation on the typer library to ^0.11.0
to get the licence checks working until the dep issues can besolved by
rstcheck.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>