💄 User template: Move account creation date up in the profile
This commit makes the account creation date appear higher up in the user
profile. In FAS 2, this information was very visible and prominent in
user profiles. It is annoying for me to scroll to the bottom of an
account for this information. This is a minor UI tweak to make it easier
to find. :D

Signed-off-by: Justin W. Flory (he/him) <jwf@redhat.com>